+++ CHANGES.md
@@ -1,8 +1,9 @@
-# Version 0.2 (2016-12-14)
+# Version 0.2 (2016-12-15)
 
   * Renamed function to load CSV file from **load** to **loadCSV**
   * Added toList function to the series
-  * Implemented CSVWriter
+  * Implemented CSV Writer
+  * Implemented apply function.
   
 
 # Version 0.1 (2016-12-13)